исправил конструктор класса exception.исправил парсер перечня товаров магазина, в лог вывод total сумм автозакупки сделал с разделением разрядов.

+++ BIZSrc/BIZ_Types.cpp
@@ -841,7 +841,7 @@ std::string GetCurrencyName(int currency)
 //-----------------------------------------------------------------------------
 //
 //-----------------------------------------------------------------------------
-std::string IntToStr(int Value)
+std::string IntToStr(long Value)
 {
     std::string res;
     char s[50];
